refactor: more general naming for frontend Zielnorm preview service
RISDEV-7876
