+++ frontend/src/services/zielnormExpressionsService.spec.ts
@@ -50,7 +50,9 @@ describe("useGetZielnormPreview", () => {
 
     vi.doMock("@/services/apiService", () => ({ useApiFetch }))
 
-    const { useGetZielnormPreview } = await import("./zielnormPreviewService")
+    const { useGetZielnormPreview } = await import(
+      "./zielnormExpressionsService"
+    )
 
     const result = useGetZielnormPreview(
       NormExpressionEli.fromString("eli/bund/bgbl-1/2021/s4/2021-03-01/1/deu"),
@@ -87,7 +89,9 @@ describe("useGetZielnormPreview", () => {
       .spyOn(window, "fetch")
       .mockResolvedValue(new Response("[]"))
 
-    const { useGetZielnormPreview } = await import("./zielnormPreviewService")
+    const { useGetZielnormPreview } = await import(
+      "./zielnormExpressionsService"
+    )
 
     const eli = ref(undefined)
     useGetZielnormPreview(eli)
@@ -100,7 +104,9 @@ describe("useGetZielnormPreview", () => {
       .spyOn(window, "fetch")
       .mockResolvedValue(new Response("[]"))
 
-    const { useGetZielnormPreview } = await import("./zielnormPreviewService")
+    const { useGetZielnormPreview } = await import(
+      "./zielnormExpressionsService"
+    )
 
     const eli = ref(
       NormExpressionEli.fromString("eli/bund/bgbl-1/2021/s4/2021-03-01/1/deu"),
@@ -155,7 +161,7 @@ describe("useCreateZielnormExpressions", () => {
     vi.doMock("@/services/apiService", () => ({ useApiFetch }))
 
     const { useCreateZielnormExpressions } = await import(
-      "./zielnormPreviewService"
+      "./zielnormExpressionsService"
     )
 
     const result = useCreateZielnormExpressions(
@@ -196,7 +202,7 @@ describe("useCreateZielnormExpressions", () => {
       .mockResolvedValue(new Response("[]"))
 
     const { useCreateZielnormExpressions } = await import(
-      "./zielnormPreviewService"
+      "./zielnormExpressionsService"
     )
 
     const expressionEli = ref(undefined)
@@ -214,7 +220,7 @@ describe("useCreateZielnormExpressions", () => {
       .mockResolvedValue(new Response("[]"))
 
     const { useCreateZielnormExpressions } = await import(
-      "./zielnormPreviewService"
+      "./zielnormExpressionsService"
     )
 
     const expressionEli = ref(
